map_a.blend now has proxies for the route, properly named tower bases and baked textures.
The proxy naming format is:
`Proxy,kind,route_id,node_id`, with the types for the `route` kind being `int,int`, which translates to a full proxy name type signature of: `String,String,int,int`. For tower bases (not proxies - tower bases proxies could also be a thing, but not in this iteration) the type signature of the name is like this: `String,int`. Example: `TowerBase,-1`, which would be a "bad" tower base. The int denotes power level, so, -1 is "bad", 0 is normal and 1 is a particularly strong or "good" tower base. Beware that TowerBase objects may have `.006`-style suffixes from copying them around and stuff - best ignore hat/cut it off for parsing - the relevant info-parts will always start at the beginning and be comma delimited, but might contain floats one day (which implies `.`), so it's best to cut the blender-thing from behind if `.` is used to do it.